refactor: make decode_logs method in ape_eth txn
